Proyecto-TA2.0-master/BackEnd/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-multipart==0.0.9
PyJWT==2.8.0
pymysql==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
pandas
openpyxl
